Move single use components into routes (#1/3)
