+++ apps/console/app/routes/_auth.signin/route.tsx
@@ -5,9 +5,9 @@ import { Badge } from "@hebo/ui/components/Badge";
 import { Button } from "@hebo/ui/components/Button";
 import { Skeleton } from "@hebo/ui/components/Skeleton";
 
-import { AuthProvider } from "~/components/auth/AuthProvider";
-import { MagicLinkSignIn } from "~/components/auth/MagicLinkSignIn";
-import { OAuthSignIn } from "~/components/auth/OAuthSignIn";
+import { AuthProvider } from "~/components/AuthProvider";
+import { MagicLinkSignIn } from "./magiclink";
+import { OAuthSignIn } from "./oauth";
 import { Logo } from "~/components/ui/Logo";
 
 export default function SignIn() {
