+++ apps/console/app/routes/_shell.user.account/route.tsx
@@ -1,0 +1,28 @@
+import { lazy } from "react";
+
+import { isStackAuthEnabled } from "~/lib/env";
+
+import { Suspense } from "react";
+
+import { Skeleton } from "@hebo/ui/components/Skeleton";
+
+import { AuthProvider } from "~/components/AuthProvider";
+
+const StackAccountSettings = lazy(() =>
+  import("@stackframe/react").then((mod) => ({ default: mod.AccountSettings })),
+);
+
+export default function Account() {
+  return isStackAuthEnabled ? (  
+    <Suspense fallback={<Skeleton count={3} className="m-4 h-20 w-full" />}>
+      <AuthProvider>
+        <StackAccountSettings />
+      </AuthProvider>
+    </Suspense>
+  ) : (
+    <>
+      <h1>Account Settings</h1>
+      <div>Not implemented for Dummy Auth</div>
+    </>
+  );
+}
